requirements.txt
astroid==2.4.2
click==7.1.2
dataclasses-json==0.5.2
Flask==1.1.2
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-enum==1.5.1
mccabe==0.6.1
mypy-extensions==0.4.3
paho-mqtt==1.5.1
pylint==2.6.0
six==1.15.0
stringcase==1.2.0
toml==0.10.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
Werkzeug==1.0.1
wrapt==1.12.1
